Fixed PR comments and refactore extended LUT API
Also updated and fixed Unit Tests

+++ base/common/DateLUTImpl.h
@@ -332,8 +332,8 @@ public:
         if (offset_is_whole_number_of_hours_everytime)
             return (toSecondsSinceTheDayStart(t) / 60) % 60;
 
-        UInt32 date = find(t).date;
-        return (UInt32(t) - date) / 60 % 60;
+        Int64 date = find(t).date; // TODO: validate for negative t values
+        return static_cast<unsigned>(Int64(t) - date) / 60 % 60;
     }
 
     inline time_t toStartOfMinute(time_t t) const { return t / 60 * 60; }
@@ -346,8 +346,8 @@ public:
         if (offset_is_whole_number_of_hours_everytime)
             return t / 3600 * 3600;
 
-        UInt32 date = find(t).date;
-        return date + (UInt32(t) - date) / 3600 * 3600;
+        Int64 date = find(t).date;
+        return date + static_cast<UInt32>(Int64(t) - date) / 3600 * 3600;
     }
 
     /** Number of calendar day since the beginning of UNIX epoch (1970-01-01 is zero)
@@ -731,7 +731,7 @@ public:
         if (time_offset >= lut[index].time_at_offset_change * 900)
             time_offset -= lut[index].amount_of_offset_change * 900;
 
-        UInt32 res = lut[index].date + time_offset;
+        Int64 res = lut[index].date + time_offset;
 
         if (unlikely(res > date_lut_max))
             return 0;
@@ -781,7 +781,7 @@ public:
     {
         const Values & values = find(t);
         return
-              toSecond(t)
+              static_cast<UInt64>(toSecond(t))
             + toMinute(t) * 100
             + toHour(t) * 10000
             + UInt64(values.day_of_month) * 1000000
